Refactored logic for filtering target and effect type specific buffs into buffs file
